requirements.txt
torch
torchvision
torchaudio
ms-swift
modelscope[framework]
datasets
transformers
sentencepiece